Create divideStringK.py
A string s , of length n.
An integer K where  is a factor of  n.
We can split  into n/k  substrings where each subtring t consists of a contiguous block of k characters in s. Then, use each t  to create string u such that:

The characters in u are a subsequence of the characters in t .
Any repeat occurrence of a character is removed from the string such that each character in u  occurs exactly once. In other words, if the character at some index j in t occurs at a previous index <j  in t , then do not include the character in string .